- made compatible with c++11     - removed -ansi -pedantic -Wno-long-long     - fixed narrowing in initializer lists

+++ src/geartgeo/MaterialMapFactory.cc
@@ -16,7 +16,7 @@ namespace gear {
 					 double zmin, double zmax, int nzsteps, 
 					 int coord)
   {
-    double gridA[]={xmin,xmax,nxsteps,ymin,ymax,nysteps,zmin,zmax,nzsteps};
+    double gridA[]={xmin,xmax,double(nxsteps),ymin,ymax,double(nysteps),zmin,zmax,double(nzsteps)};
     std::vector<double> gridV;
     gridV.assign(gridA,gridA+9);
